JPEG-XL uses 1 byte, not 2
Currently, the HDR in PNG proposal points out that it is different
from JPEG-XL because this would use 1 byte instead of 2. But
JPEG-XL also uses 1 byte.

The real difference is this proposal uses H.273 where JPEG-XL
cherry picked a handful of color spaces and left room for custom
values.

This commit corrects the 1 vs. 2 byte description and clarifies the
H.273 difference.

Closes #28