iCCP and iCCN specify different ICC versions
The HDR in PNG proposal currently says the only difference between
the iCCP and iCCN chunks is Latin-1 vs. UTF-8 profile names.

But this is not completely correct. The PNG spec explicitly says
iCCP is ICCv2. Although many decoders treat iCCP as ICCv4, the
iCCN chunk would be explicitly ICCv4.

This commit clarifies this distinction.

Closes #30