Remove inconsistent behavior
In the current HDR in PNG proposal, there is a line that describes
a decoder encountering a cICP chunk and an iCCP chunk would have
undefined behavior.

However, later in the same proposal the exact behavior is spelled out.

This commit removes the undefined behavior, leaving the behavior
agreed upon.

Closes #32