Explicitly state two proposed chunks
Currently, the proposal mentions the details of two chunks without
explaining that these two chunks are the proposed new chunks to
address the issue.

To people who were at the meetings, it may be clear enough. But to
an external reader, it could seem like a mind dump without clarity.

This commit explicitly states that the two chunks are new to this
proposal and how they should work.