Explicitly state two proposed chunks (#59)
* Explicitly state two proposed chunks

Currently, the proposal mentions the details of two chunks without
explaining that these two chunks are the proposed new chunks to
address the issue.

To people who were at the meetings, it may be clear enough. But to
an external reader, it could seem like a mind dump without clarity.

This commit explicitly states that the two chunks are new to this
proposal and how they should work.

* Separate concerns

Currently, two separate concerns are in the same paragraph.

This commit separates those two concerns into two paragraphs.

Co-authored-by: Pierre-Anthony Lemieux <pal@sandflow.com>

* Fix typo

The word "an" was typo'd as "en".

This commit fixes the typo.

Co-authored-by: Pierre-Anthony Lemieux <pal@sandflow.com>

* Add chunk order information at each chunk definition

Wording along the lines of "both of these chunks..." is not ideal.
A reader could rightfully jump to the chunk they are interested in
and skip that wording.

This commit adds shared wording to both chunk definitions.

Co-authored-by: Pierre-Anthony Lemieux <pal@sandflow.com>